correction courant LED et affichage interface

+++ MAX30102.h
@@ -45,7 +45,7 @@ uint8_t readRegistre(uint8_t adresse);
 void writeRegistre(uint8_t data, uint8_t adresse);
 void MAX30102_config();
 void readFIFO(uint32_t *red_LED, uint32_t *ir_LED, uint16_t compteur);
-void changeLED_red (short int ledAmp_red);
-void changeLED_IR (short int ledAmp_IR);
+void changeLED_red (uint8_t ledAmp_red);
+void changeLED_IR (uint8_t ledAmp_IR);
 
 /* [] END OF FILE */
